@@ -11,7 +11,7 @@
 		}
 	],
 	"require": {
-		"php": ">=7.0|>=8.0"
+		"php": ">=7.3|>=8.0"
 	},
 	"autoload": {
 		"psr-4": {
