@@ -2,7 +2,6 @@
 
 namespace JiJiHoHoCoCo\IchiORM\Database\Connections;
 
-use PDO;
 class SQLiteConnection extends Connection{
 
 
@@ -12,7 +11,7 @@ class SQLiteConnection extends Connection{
 			throw new \Exception("You must add database name and host for SQL Lite Database Connection", 1);
 		}
 
-		$dsn=$config['driver'] . ':' . $config['dbname'];
+		$dsn = $config['driver'] . ':' . $config['dbname'];
 
 		return $dsn;
 	}
