@@ -4,13 +4,13 @@ namespace JiJiHoHoCoCo\IchiORM\Pagination;
 
 class Paginate{
 
-	private $per_page,$pageCheck,$current_page,$start;
+	private $per_page, $pageCheck, $current_page, $start;
 
 	public function setPaginateData(int $per_page){
-		$this->per_page=$per_page;
-		$this->pageCheck=pageCheck();
-		$this->current_page=$this->pageCheck ? intval($_GET['page']) : 1;
-		$this->start=($this->current_page>1) ? ($this->per_page*($this->current_page-1)) : 0 ;
+		$this->per_page = $per_page;
+		$this->pageCheck = pageCheck();
+		$this->current_page = $this->pageCheck ? intval($_GET['page']) : 1;
+		$this->start = ($this->current_page>1) ? ($this->per_page*($this->current_page-1)) : 0 ;
 	}
 
 	public function getStart(){
@@ -18,14 +18,14 @@ class Paginate{
 	}
 
 	public function paginate($total,$objectArray){
-		$total_pages=ceil($total/$this->per_page);
-		$next_page=$this->current_page+1;
-		$previous_page=$this->pageCheck && $_GET['page']-1>=1 ? $_GET['page']-1 : NULL ;
-		$from=$this->start+1;
+		$total_pages = ceil($total/$this->per_page);
+		$next_page = $this->current_page+1;
+		$previous_page = $this->pageCheck && $_GET['page']-1>=1 ? $_GET['page']-1 : NULL ;
+		$from = $this->start+1;
 
-		$domainName=getDomainName();
-		$totalPerPage=count($objectArray);
-		$to=($from+$totalPerPage)-1;
+		$domainName = getDomainName();
+		$totalPerPage = count($objectArray);
+		$to = ($from+$totalPerPage)-1;
 
 		return [
 			'current_page' => $this->current_page,
